requirements.txt
openai
python-dotenv
edge_tts
pydub
speechrecognition 